% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create the exposure cohorts}
\usage{
createCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortToCreateFile,
  packageName = "VaTools",
  cohortTable,
  oracleTempSchema,
  outputFolder,
  databaseId,
  generateInclusionStats = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in
the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your vocabulary tables in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward
slashes (/)}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{tablePrefix}{A prefix to be used for all table names created for this study.}

\item{indicationId}{A string denoting the indicationId for which the exposure cohorts
should be created.}
}
\description{
Create the exposure cohorts
}
\details{
This function will create the exposure cohorts following the definitions included in this package.
}
