% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStandardCdmSchema.R
\name{createStandardCdmSchema}
\alias{createStandardCdmSchema}
\title{Create standard CDM schema}
\usage{
createStandardCdmSchema(
  connectionDetails,
  database,
  startingSchema,
  destinationSchema = "OMPV5",
  tableNamePrefix = "OMOPV5_"
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details}

\item{database}{Database in which to create the new views (e.g. "ORD_Researcher_xyz")}

\item{startingSchema}{Schema containing provisioned CDM views (e.g. "src")}

\item{destinationSchema}{Destination schema for compliant views}

\item{tableNamePrefix}{Prefix of non-standard view names (e.g. "omopv5_")}
}
\value{
invisible NULL
}
\description{
Create standard CDM schema
}
