% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortNames.R
\name{getCohortNames}
\alias{getCohortNames}
\title{getCohortNames}
\usage{
getCohortNames(
  connectionDetails,
  database,
  schema = "src",
  table = "omopv5_person",
  columnName = "CohortName"
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connectionDetails object}

\item{database}{"ORD_Name_xyz"}

\item{schema}{"src"}

\item{table}{"omopv5_person"}
}
\value{
Distinct values of CohortName (character vector)
}
\description{
Get distinct values of CohortName from person table
}
\details{
Due to the way VINCI provisions cohorts, there may be duplicate rows for a person_id in the CDM person table.
This helper function lists the distinct values of the non-standard column `CohortName`.
}
