% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaTools.R
\docType{package}
\name{VaTools-package}
\alias{VaTools}
\alias{VaTools-package}
\title{VaTools: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
